rockchip: rk3568-adakta-smarc: enable npu regulator
The power-domain driver in Linux does not know what regulator is
supplying power to a specific power-domain. This prevent use of NPU with
vendor kernel because vdd_npu is disabled.

Change vdd_npu to use always-on/boot-on to enable the regulator at boot
and set initial 0v9 voltage, the recommended voltage in soc datasheet.